Update useExpanded Doc to add missing property
There is a getToggleRowExpandedProps that exists as a Instance Property of useTable that is used throughout the [examples](https://github.com/tannerlinsley/react-table/search?q=getToggleRowExpandedProps) on the project, but it is not listed on the docs. This adds it in the property alongside the other useExpanded properties  with a brief description on what it does.